+++ packages/jest-core/src/SearchSource.ts
@@ -83,9 +83,19 @@ const globsToMatcher = (globs: Array<Config.Glob>) => {
   };
 };
 
-const regexToMatcher = (testRegex: Config.ProjectConfig['testRegex']) => (
-  path: Config.Path,
-) => testRegex.some(testRegex => new RegExp(testRegex).test(path));
+const regexToMatcher = (testRegex: Config.ProjectConfig['testRegex']) => {
+  const regexes = testRegex.map(testRegex => new RegExp(testRegex));
+
+  return (path: Config.Path) =>
+    regexes.some(regex => {
+      const result = regex.test(path);
+
+      // prevent stateful regexes from breaking, just in case
+      regex.lastIndex = 0;
+
+      return result;
+    });
+};
 
 const toTests = (context: Context, tests: Array<Config.Path>) =>
   tests.map(path => ({